+++ prompts/ai_report_prompts.py
